split hyperparameters_test.py
